backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.1.1

# Supabase (compatible versions)
supabase==2.9.0
httpx==0.27.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Pydantic - MUST match CrewAI requirements
pydantic>=2.11.9
pydantic-settings>=2.10.1
anyio>=4.7.0

# Existing dependencies (CrewAI integration)
# Your existing crewai dependencies are already installed
